Adds support for ID value in slots information
Slots are the values that are parsed at Alexas cloud service. A synonym is insertetd into the slot if it was used in the parsed utterance. 
This patch ads the posibillity to use the ID field of  the slot to get a consistent value over all the synonymes for a particular value. This is also usefull if you have multi language apps.